refactor(datashare-tasks): create namespace and custom search attributes at app startup
