+++ datashare-app/src/main/java/org/icij/datashare/tasks/TaskManagerTemporal.java
@@ -2,8 +2,8 @@ package org.icij.datashare.tasks;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import io.temporal.client.WorkflowClient;
 import org.icij.datashare.PropertiesProvider;
+import org.icij.datashare.asynctasks.temporal.TemporalInterlocutor;
 
 
 @Singleton
@@ -11,7 +11,7 @@ public class TaskManagerTemporal extends org.icij.datashare.asynctasks.TaskManag
     implements DatashareTaskManager {
 
     @Inject
-    public TaskManagerTemporal(WorkflowClient client, PropertiesProvider propertiesProvider) {
-        super(client, Utils.getRoutingStrategy(propertiesProvider));
+    public TaskManagerTemporal(TemporalInterlocutor temporal, PropertiesProvider propertiesProvider) {
+        super(temporal, Utils.getRoutingStrategy(propertiesProvider));
     }
 }
