+++ datashare-tasks/src/test/java/org/icij/datashare/asynctasks/TaskManagerTemporalIntTest.java
@@ -8,9 +8,6 @@ import static org.icij.datashare.asynctasks.Task.State.ERROR;
 import static org.icij.datashare.asynctasks.Task.State.RUNNING;
 import static org.icij.datashare.asynctasks.TaskManagerTemporal.DEFAULT_NAMESPACE;
 import static org.icij.datashare.asynctasks.TaskManagerTemporal.WORKFLOWS_DEFAULT;
-import static org.icij.datashare.asynctasks.TaskManagerTemporal.buildClient;
-import static org.icij.datashare.asynctasks.TaskManagerTemporal.deleteNamespace;
-import static org.icij.datashare.asynctasks.TaskManagerTemporal.setupNamespace;
 import static org.icij.datashare.asynctasks.temporal.TemporalHelper.activityFactory;
 import static org.icij.datashare.asynctasks.temporal.TemporalHelper.createTemporalWorkerFactory;
 import static org.junit.Assert.assertThrows;
@@ -38,6 +35,7 @@ import org.icij.datashare.asynctasks.temporal.FailingWorkflowImpl;
 import org.icij.datashare.asynctasks.temporal.HelloWorldActivityImpl;
 import org.icij.datashare.asynctasks.temporal.HelloWorldWorkflowImpl;
 import org.icij.datashare.asynctasks.temporal.TemporalHelper;
+import org.icij.datashare.asynctasks.temporal.TemporalInterlocutor;
 import org.icij.datashare.tasks.RoutingStrategy;
 import org.icij.datashare.user.User;
 import org.junit.Before;
@@ -52,7 +50,7 @@ public class TaskManagerTemporalIntTest {
     @Mock
     private TaskFactory taskFactory;
 
-    private static WorkflowClient client;
+    private static TemporalInterlocutor temporal;
 
     private static TaskManagerTemporal taskManager;
 
@@ -60,16 +58,16 @@ public class TaskManagerTemporalIntTest {
 
 
     @BeforeClass
-    public static void setUpClass() {
-        client = buildClient(EnvUtils.resolve("temporalTarget", "temporal:7233"), DEFAULT_NAMESPACE);
-        taskManager = new TaskManagerTemporal(client, RoutingStrategy.UNIQUE);
+    public static void setUpClass() throws InterruptedException {
+        temporal = new TemporalInterlocutor(EnvUtils.resolve("temporalTarget", "temporal:7233"), DEFAULT_NAMESPACE);
+        taskManager = new TaskManagerTemporal(temporal, RoutingStrategy.UNIQUE);
     }
 
     @Before
     public void setUp() throws IOException, InterruptedException {
         mocks = openMocks(this);
         try {
-            deleteNamespace(client, Duration.ofSeconds(5));
+            temporal.deleteNamespace(Duration.ofSeconds(5));
         } catch (StatusRuntimeException ex) {
             if (!ex.getStatus().getCode().equals(Status.Code.NOT_FOUND)) {
                 throw ex;
@@ -79,14 +77,14 @@ public class TaskManagerTemporalIntTest {
             new TemporalHelper.RegisteredWorkflow(
                 HelloWorldWorkflowImpl.class,
                 WORKFLOWS_DEFAULT,
-                List.of(new TemporalHelper.RegisteredActivity(activityFactory(HelloWorldActivityImpl.class, taskFactory, client, 1.0d), WORKFLOWS_DEFAULT))),
+                List.of(new TemporalHelper.RegisteredActivity(activityFactory(HelloWorldActivityImpl.class, taskFactory, temporal.client, 1.0d), WORKFLOWS_DEFAULT))),
             new TemporalHelper.RegisteredWorkflow(
                 FailingWorkflowImpl.class,
                 WORKFLOWS_DEFAULT,
-                List.of(new TemporalHelper.RegisteredActivity(activityFactory(FailingActivityImpl.class, taskFactory, client, 1.0d), WORKFLOWS_DEFAULT)))
+                List.of(new TemporalHelper.RegisteredActivity(activityFactory(FailingActivityImpl.class, taskFactory, temporal.client, 1.0d), WORKFLOWS_DEFAULT)))
         );
 
-        setupNamespace(client, Duration.ofSeconds(5));
+        temporal.setupNamespace(Duration.ofSeconds(5));
         taskManager.clear();
         Thread.sleep(2000); // Sleep to allow custom attribute creation propagation refresh rate is 0.1s
     }
@@ -199,7 +197,7 @@ public class TaskManagerTemporalIntTest {
 
     @Test(timeout = 10000)
     public void test_get_task_result() throws IOException {
-        try (TemporalHelper.CloseableWorkerFactoryHandle ignored = testCloseableWorkerFactory(client)) {
+        try (TemporalHelper.CloseableWorkerFactoryHandle ignored = testCloseableWorkerFactory(temporal.client)) {
             Task<String> task = new Task<>("hello-world", User.local(), Map.of("name", "world"));
 
             taskManager.startTask(task);
@@ -213,7 +211,7 @@ public class TaskManagerTemporalIntTest {
 
     @Test(timeout = 20000)
     public void test_task_error() throws IOException, InterruptedException {
-        try (TemporalHelper.CloseableWorkerFactoryHandle ignored = testCloseableWorkerFactory(client)) {
+        try (TemporalHelper.CloseableWorkerFactoryHandle ignored = testCloseableWorkerFactory(temporal.client)) {
             Task<String> task = new Task<>("failing", User.local(), Map.of());
 
             taskManager.startTask(task);
@@ -251,7 +249,7 @@ public class TaskManagerTemporalIntTest {
     @Ignore("keeping this one for manual test as it can take very long to complete")
     @Test
     public void test_clear_done_tasks() throws Exception {
-        try (TemporalHelper.CloseableWorkerFactoryHandle ignored = testCloseableWorkerFactory(client)) {
+        try (TemporalHelper.CloseableWorkerFactoryHandle ignored = testCloseableWorkerFactory(temporal.client)) {
             Task<String> task = new Task<>("hello-world", User.local(), Map.of("key", "value"));
             taskManager.startTask(task);
             assertThat(taskManager.awaitTermination(2, TimeUnit.SECONDS));
@@ -266,7 +264,7 @@ public class TaskManagerTemporalIntTest {
     @Ignore("keeping this one for manual test as it can take very long to complete")
     @Test
     public void test_clear_done_tasks_with_args_filter() throws IOException, InterruptedException {
-        try (TemporalHelper.CloseableWorkerFactoryHandle ignored = testCloseableWorkerFactory(client)) {
+        try (TemporalHelper.CloseableWorkerFactoryHandle ignored = testCloseableWorkerFactory(temporal.client)) {
             Task<String> first = new Task<>("hello-world", User.local(), Map.of("key", "value"));
             Task<String> second = new Task<>("hello-world", User.local(), Map.of("otherKey", "otherValue"));
             taskManager.startTask(first);
